+++ test/fixtures/cookbooks/test/recipes/configure.rb
@@ -169,3 +169,27 @@ grafana_dashboard 'sample-dashboard-folder' do
   )
   action [:create, :update]
 end
+
+grafana_dashboard_template 'sample-dashboard-template' do
+  auth_proxy_header auth_header
+
+  organization 'Sous-Chefs'
+  folder 'StayOrganized2'
+
+  template_source 'dashboards/sample-dashboard-template.json.erb'
+  template_cookbook template[:cookbook]
+
+  template_vars(
+    random_walk: 'injected value'
+  )
+
+  action [:create, :update]
+end
+
+grafana_dashboards_backup 'Backup Dashboards to File' do
+  auth_proxy_header auth_header
+end
+
+grafana_datasources_backup 'Backup Datasources to File' do
+  auth_proxy_header auth_header
+end