+++ test/fixtures/cookbooks/test/recipes/plugins.rb
@@ -11,6 +11,10 @@ grafana_plugin 'grafana-clock-panel' do
   action :install
 end
 
+grafana_plugin 'yesoreyeram-boomtable-panel' do
+  plugin_url 'https://github.com/axelrtgs/yesoreyeram-boomtable-panel/archive/feature/sorting.zip'
+end
+
 # Tests are failing as the server has not fully become available when tests run
 chef_sleep 'Sleep so inspec tests pass' do
   seconds 25